requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dateparser==1.1.7
fastapi==0.90.0
fonttools==4.38.0
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.3
mplfinance==0.12.9b7
multidict==6.0.4
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pydantic==1.10.4
pyparsing==3.0.9
python-binance==1.0.16
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
six==1.16.0
sniffio==1.3.0
starlette==0.23.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
yarl==1.8.2
